bartok: delete command works properly (for project/service), local storage tweaks

+++ bartok/server/manager/instanceKill.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 
-const dirname = `${__dirname}/../__services`;
+const dirname = `${__dirname}/../../__services`;
 
 var deleteFolderRecursive = function (path) {
 	if (fs.existsSync(path)) {
@@ -13,19 +13,34 @@ var deleteFolderRecursive = function (path) {
 			}
 		});
 		fs.rmdirSync(path);
+	} else {
+		console.error();
+		console.error(`WARNING: unfound: ${path}!`);
+		console.error();
 	}
 };
 
 function instanceKill(services, callback) {
-	// TODO: should also delete file(s) for service
-	// next is a bad way to do that
-	if(services.length > 1){
-		deleteFolderRecursive(dirname);
-	}
+
 	(async () => {
 		//console.log({ servicesToKill: services.map(x => x.id) })
+		//console.log( services )
+
 		for(var i=0, len=services.length; i < len; i++){
-			await services[i].instance.kill()
+			const instance = services[i].instance;
+			if(!instance){
+				console.error();
+				console.error(`Why does this service not have an instance?`);
+				console.error(services[i]);
+				console.error();
+				continue;
+			}
+			await services[i].instance.kill();
+			// TODO: should also delete file(s) for service
+			// next is a bad way to do that?
+			const dirToDelete = `${dirname}/.${services[i].name.replace(/\s/g, '_')}.service`;
+			//console.log({ dirToDelete });
+			deleteFolderRecursive(dirToDelete);
 		}
 		callback && callback();
 	})();
